Extend Duration to years and weeks
I'm currently working on unifying all Duration parsing. While it's
arguably rarely useful to use y or w, I'm pretty sure that we have
users doing it (and we have separate parsing code still in place
supporting it (and I'm working on removing that other parsing code)).

I'd hate to break people's expressions and configs just because we
think 1w is too long for a range...