Return meaningful error instead of EOF during text parsing
So far, we are just passing through the io.EOF error created when
trying to read from the input stream. That io.EOF error doesn't tell
the caller the position where it has happened, and worse, it could
easily misunderstood as a signal for the legitimate end of the input
stream.