route: Allow per-request custom context injection
**Motivation**

Currently, Prometheus's web API knows of two different contexts:

- The main Prometheus server context which is passed into the web API
  and through to the query layer.
- The per-request context which is generated by the route.Router. This
  is currently only used for storing and retrieving URL-path parameters,
  and it's not possible to influence this context at the top of the
  chain.

Since merging of contexts is not a concept, we ultimately want the
router to support a per-request context which:

a) can be derived from the main server context,
b) can be decorated with custom values/timeouts/etc. per-request

The concrete motivation for b) is for Weaveworks Frankenstein to be able
to extract multi-tenancy authentication headers and store user
information in the context. This will then be passed on into the query
layers and decide what data a user gets to see.

Next step: update Prometheus to make use of this, get rid of the context
in the web API itself. This will make the API package reusable by
Frankenstein.