Adjust the MetricNameRE to reflect the actual behavior
Note that the validation in
https://github.com/prometheus/common/blob/master/model/metric.go#L93
does not actually use MetricNameRE. Right now, the regexp is more
documentation than anything else, although some might use it (which is
one of the reasons to keep it around).

In any case, the regexp should match exactly the same strings that are
matched by the IsValidMetricName function in
https://github.com/prometheus/common/blob/master/model/metric.go#L93
matches.