Don't panic when route.Param has no binding.
I've adopted route/common for a Prom related project, and had panics
due to a middleware attempting to grab a Param that had no binding in
the URL path in some cases.

I do like the safety aspect of panicing here, but strings have a
useful zero-value of "", which seems like an appropriate, and better
choice for this functionality.

(This will aide in
https://github.com/prometheus/pushgateway/issues/149 as some paths
don't have a *labels binding, but probably want to share some of the
same processing of labels code.)

Signed-off-by: Andrew G10z <agwozdziewycz@salesforce.com>