Allow TLS renegotiation
This patch is to address an interop issue against Microsoft IIS. If, for example, one tries to use client certificates against such a server, there will always be a TLS renegotiation error. This has been addressed a while ago in Go by adding the 'Renegotiation' parameter.

As an alternative, this could be made configurable.