Make `version.NewCollector` more idiomatic
Discussing a documentation request in client_golang (
https://github.com/prometheus/client_golang/issues/693 ), I realized
that the usage here in the `version` package is not as idiomatic as it
could be. Returning a `GaugeVec`, the metric value could easily be
changed. This could be done with a `NewConstMetric`, but I think the
currently best way for this is a `GaugeFunc`. It's more compact and
avoids misaligning the label values. Plus, it returns just a
`Collector`, so no misunderstanding what to do with it (like adding
more elements to the `GaugeVec` or changing the value).

Signed-off-by: beorn7 <beorn@grafana.com>