Implement EncoderCreator in expfmt
Main reason: allow to use custom encoders in Prometheus.
Changes are backwards compatible.

If one would like to implement custom formatting of metrics,
it wasn't possible. There was no way, because both evaluation
of which format to use and list of encoders were hardcoded
in the code.

Solution is to create a (configurable) instance that would accept the
additional encoder implementations (see EncoderImplementation struct).

Signed-off-by: Mateusz Warzyński <mathew.warzynski@gmail.com>