Change Secret to SecretLoader
SecretLoader implements a number of methods to allow for retrieving a
secret from either a file or from memory. In order to avoid having to
change existing configurations, the code is careful to load the filename
specified in the corresponding _file field (e.g. password,
password_field). It's assumed that a configuraiton specifying both the
secret and the filename is invalid.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>