POC: Serve a nice HTML page
One of the main reasons we hear for defaulting to the text format is
that "look at the raw metrics in a browser" is a common use case.

Browsers send accept headers, so we can explicitly take this use case
into account, and render an HTML representation of the metric family.

For the purposes of the PoC, this simply wraps the text format. To
really make things nicer, we can and should render the metrics in a more
appealing format, with labels in a table, and possibly a bar chart for
histograms.

Signed-off-by: Matthias Rampke <matthias@prometheus.io>