Don't always fetch a OAuth2 token, if the secret from a file didn't change
When configuring a OAuth2 client using the `client_secret_file`
configuration option, the original intent of the code was to *not*
forcefully refetch a token if the content of the file didn't change.

However, both the test and the implementation had 2 typos and a new
token token was fetched on every request through the HTTP client.

Signed-off-by: Jonathan Ballet <jon@multani.info>