Remove secret file existence check in Validate for headers
This commit removes the check for the existence of the secret file for
headers in the `Validate` function. The check is not valid for relative
paths since `Validate` is called before `SetDirectory`, which is
responsible for setting up the environment. This change aligns with the
handling of other secret files, which are not checked during config
validation.

Signed-off-by: Julien <roidelapluie@o11y.eu>