promslog: Use the default timezone (again)
As per the dev-summit discussion, we want to use the default timezone
as selected by slog after all rather than a hardcoded UTC.

(This was accidentally introduced earlier when moving from gokit to
slog, but then switched back to hardcoded UTC in PR #735.)

Environments that have configured UTC anyway won't see any difference,
but those with other timezones configured will see the configured
timezone for the timestamps in their logs. This is in line with how
slog behaves by default.

Example timestamp _before_ this change and also _after_ this change
with UTC configured in the environment via a suitable way (e.g.
setting `TZ=UTC` on Linux):

    time=2025-04-10T12:00:38.179Z

Example timestamp _after_ this change with location Europe/Berlin,
i.e. timezone is CET or CEST:

    time=2025-04-10T14:00:03.120+02:00

Note that the precise delta to UTC is in the timestamp. So this change
will also be transparent to the usual logs processing tools.

Signed-off-by: beorn7 <beorn@grafana.com>