Add support for target scraping with SPIFFE
This is the first step toward
https://github.com/prometheus/common/issues/702

The feature is not available until the calling code invokes the
WithSpiffeSourceFactory HTTPClientOptions, so no behaviour is yet changed.
Use of that option will be the subject of a future change to
https://github.com/prometheus/prometheus

SPIFFE replaces the usual peer X.509 certificate verification
algorithm with its own that checks SPIFFE IDs in URI SANs and checks
against different sets of trust roots for different SPIFFE trust domains.
Accordingly, none of the other TLS configuration parameters are applicable
when SPIFFE is configured and vice versa and this mutual exclusivity is
enforced in tls_config.

There are two ways to set which SPIFFE ID should be expected from a
scrape endpoint that is using HTTPS and SPIFFE: in the tls_config, and
per-request. The former would be expected mainly on static scrape configs
that define a single (perhaps replicated) scrape endpoint. For scrape
configs with target discovery it is expected that different endpoints
would present certificates with different SPIFFE IDs and so the
per-request version would be used. It is intended that the peer's
expected SPIFFE ID should come from a new special label __spiffe_id__
which would be populated by target discovery. That too will be part of
the next change. For the purposes of this library, the per-request peer
SPIFFE ID is supplied in the Context accompanying the Request.

Signed-off-by: Kim Vandry <vandry@TZoNE.ORG>